Rename aunit_cb/using_aunit_with_restricted_run-time_libraries.rst
The name of this file is very long, and this limits on Windows hosts
the location where the documentation can be unpacked, due to the
maximum path limitation we have there. This commit reduces
the likeliness of this problem happening, by shortening the length
of that file.

Tested by rebuilding the documentation in PDF format.

TN: TC18-012